Move Spatial to data package
